# openapi.yaml
openapi: "3.0.0"
info:
  title: "Not Chocolatine API"
  version: "1.0.0"
servers:
  - url: "http://localhost:8080"
paths:
...
    /users:
        get:
            summary: "Get all users"
            operationId: "getUsers"
            responses:
                '200':
                    description: "Users found"
                    content:
                        application/json:
                            schema:
                                type: "array"
                                items:
                                    type: "object"
                                    properties:
                                        userId:
                                            type: "integer"
                                        mail:
                                            type: "string"
                                        password:
                                            type: "string"
                '404':
                    description: "Users not found"
                '500':
                    description: "Internal server error"
    /users/connect:
        get:
            summary: "Get user by mail and password"
            operationId: "getUserByMailAndPassword"
            parameters:
                - name: "mail"
                  in: "query"
                  required: true
                  schema:
                      type: "string"
                - name: "password"
                  in: "query"
                  required: true
                  schema:
                      type: "string"
            responses:
                '200':
                    description: "User found"
                    content:
                        application/json:
                            schema:
                                type: "object"
                                properties:
                                    userId:
                                        type: "integer"
                                    mail:
                                        type: "string"
                                    password:
                                        type: "string"
                '404':
                    description: "User not found"
                '500':
                    description: "Internal server error"
    /users/newUser: 
        post:
            summary: "Add user"
            operationId: "addUser"
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: "object"
                            properties:
                                mail:
                                    type: "string"
                                password:
                                    type: "string"
            responses:
                '201':
                    description: "User added"
                '400':
                    description: "Bad request"
                '500':
                    description: "Internal server error"
    /users/{userId}:
        delete:
            summary: "Delete user"
            operationId: "deleteUser"
            parameters:
                - name: "userId"
                  in: "query"
                  required: true
                  schema:
                      type: "string"
            responses:
                '200':
                    description: "User deleted"
                '404':
                    description: "User not found"
                '500':
                    description: "Internal server error"
...